Updating examples
