+++ dev/examples/Shared-layout-lightbox-crossfade.tsx
@@ -1,29 +1,13 @@
 import * as React from "react"
 import { CSSProperties, useState } from "react"
-import {
-    motion,
-    AnimateSharedLayout,
-    AnimatePresence,
-    useIsPresent,
-    Transition,
-} from "@framer"
+import { motion, AnimatePresence } from "@framer"
 
 /**
  * This demonstrates children with layoutId animating
  * back to their origin components
  */
 
-const params = new URLSearchParams(window.location.search)
-const instant = params.get("instant") || false
-const partialEase = params.get("partial-ease") || false
-const type = params.get("type") || "switch"
-let transition: Transition = instant ? { type: false } : { duration: 3 }
-if (partialEase) {
-    transition = {
-        duration: 0.15,
-        ease: () => 0.1,
-    }
-}
+const transition = { type: "spring", stiffness: 500, damping: 30 }
 
 function Gallery({ items, setIndex }) {
     return (
@@ -34,15 +18,10 @@ function Gallery({ items, setIndex }) {
                     onClick={() => setIndex(i)}
                     style={{ ...item, backgroundColor: color, borderRadius: 0 }}
                     layoutId={color}
-                    transition={transition}
-                    id={i === 0 ? `item-parent` : undefined}
+                    //transition={{ duration: 5 }}
+                    id={i === 0 && "list-red"}
                 >
-                    <motion.div
-                        style={child}
-                        id={i === 0 ? `item-child` : undefined}
-                        layoutId={`child-${color}`}
-                        transition={transition}
-                    />
+                    <motion.div style={child} layoutId={`child-${color}`} />
                 </motion.li>
             ))}
         </ul>
@@ -50,38 +29,33 @@ function Gallery({ items, setIndex }) {
 }
 
 function SingleImage({ color, setIndex }) {
-    const isPresent = useIsPresent()
-
     return (
         <>
             <motion.div
                 initial={{ opacity: 0 }}
                 animate={{ opacity: 1 }}
                 exit={{ opacity: 0 }}
-                style={{
-                    ...overlay,
-                    pointerEvents: isPresent ? "auto" : "none",
-                }}
+                style={overlay}
                 id="overlay"
-                transition={transition}
+                transition={{ duration: 2 }}
                 onClick={() => setIndex(false)}
             />
             <div style={singleImageContainer}>
                 <motion.div
-                    id="parent"
+                    id="color"
                     layoutId={color}
                     style={{
                         ...singleImage,
                         backgroundColor: "#fff",
                         borderRadius: 50,
                     }}
-                    transition={transition}
+                    transition={{ duration: 2 }}
                 >
                     <motion.div
                         style={{ ...child, backgroundColor: "black" }}
                         id="child"
                         layoutId={`child-${color}`}
-                        transition={transition}
+                        transition={{ duration: 2 }}
                     />
                 </motion.div>
             </div>
@@ -89,17 +63,8 @@ function SingleImage({ color, setIndex }) {
     )
 }
 
-export function Component() {
+export function App() {
     const [index, setIndex] = useState<false | number>(false)
-
-    if (partialEase) {
-        if (index === 0) {
-            transition.ease = () => 0.1
-        } else {
-            transition.ease = (t: number) => (t === 1 ? 1 : 0.9)
-        }
-    }
-
     return (
         <div style={background}>
             <Gallery items={colors} setIndex={setIndex} />
@@ -109,14 +74,6 @@ export function Component() {
                 )}
             </AnimatePresence>
         </div>
-    )
-}
-
-export function App() {
-    return (
-        <AnimateSharedLayout type={type}>
-            <Component />
-        </AnimateSharedLayout>
     )
 }
 
@@ -195,4 +152,5 @@ const child = {
     borderRadius: 25,
     backgroundColor: "white",
     opacity: 0.5,
+    //opacity: 0.5,
 }
