+++ dev/examples/Shared-layout-lists.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { useState } from "react"
-import { motion, AnimateSharedLayout, AnimatePresence } from "@framer"
+import { motion, AnimatePresence } from "@framer"
 
 /**
  * This is a stress test of snapshotting ability as components
@@ -67,20 +67,18 @@ export const App = () => {
                 alignItems: "center",
             }}
         >
-            <AnimateSharedLayout type="crossfade" transition={{ duration: 2 }}>
-                <div style={styles.container}>
-                    <List
-                        list={lists[0]}
-                        onItemClick={(id) => moveItem(id, 1, lists, setLists)}
-                        backgroundColor="#ff3366"
-                    />
-                    <List
-                        list={lists[1]}
-                        onItemClick={(id) => moveItem(id, 0, lists, setLists)}
-                        backgroundColor="#0099ff"
-                    />
-                </div>
-            </AnimateSharedLayout>
+            <div style={styles.container}>
+                <List
+                    list={lists[0]}
+                    onItemClick={(id) => moveItem(id, 1, lists, setLists)}
+                    backgroundColor="#ff3366"
+                />
+                <List
+                    list={lists[1]}
+                    onItemClick={(id) => moveItem(id, 0, lists, setLists)}
+                    backgroundColor="#0099ff"
+                />
+            </div>
         </div>
     )
 }
