+++ dev/examples/Shared-layout-reparenting.tsx
@@ -1,6 +1,12 @@
 import * as React from "react"
-import { motion, useCycle, AnimateSharedLayout } from "@framer"
+import { motion, useCycle } from "@framer"
 import styled from "styled-components"
+
+/**
+ * This example demonstrates using shared layout
+ * to animate between two sets of two components with a different
+ * hierarchy
+ */
 
 const Container = styled.div`
     width: 200px;
@@ -54,12 +60,7 @@ const Big = styled(motion.div)`
 
 const Child = () => {
     return (
-        <Big
-            layoutId="big"
-            transformTemplate={(_, gen) =>
-                `translateX(-50%) translateY(-50%) ${gen}`
-            }
-        >
+        <Big layoutId="big">
             <Small layoutId="small" />
         </Big>
     )
@@ -68,13 +69,7 @@ const Child = () => {
 const Sibling = () => {
     return (
         <>
-            <Big
-                layoutId="big"
-                purple
-                transformTemplate={(_, gen) =>
-                    `translateX(-50%) translateY(-50%) ${gen}`
-                }
-            />
+            <Big layoutId="big" purple />
             <Small layoutId="small" purple />
         </>
     )
@@ -85,9 +80,7 @@ export const App = () => {
 
     return (
         <Container onClick={() => toggleOn()}>
-            <AnimateSharedLayout>
-                {isOn ? <Child /> : <Sibling />}
-            </AnimateSharedLayout>
+            {isOn ? <Child /> : <Sibling />}
         </Container>
     )
 }
