+++ src/projection/shared/stack.ts
@@ -60,4 +60,8 @@ export class NodeStack {
              */
         }
     }
+
+    exitAnimationComplete() {
+        this.members.forEach((node) => node.options.onExitComplete?.())
+    }
 }
